refactor(P060): collapse client-tools/schemas.ts into tools.ts
The tool-name registry (CLIENT_TOOL_NAMES, ClientToolName,
isClientToolName) lived next to the LLM_STATIC_TOOLS map but in a
separate file for no real reason. Merging both into tools.ts puts every
LLM-tool decision (which schema, which name, the runtime guard) in one
place. Adding a tool now touches one file in the adapter (tools.ts) +
one switch arm in factory.ts.
