+++ copilot/src/lib/embed-bridge-adapters/client-tools/index.ts
@@ -4,6 +4,5 @@ export type { FinalisationAction, FinalisationToolMap } from './finalisation'
 export { FINALISATION_ACTION, FINALISATION_TOOL, withFinalisationTool } from './finalisation'
 export type { MiddlewareContext, ToolMiddleware } from './middleware'
 export { composeMiddleware } from './middleware'
-export type { ClientToolName } from './schemas'
-export { CLIENT_TOOL_NAMES, isClientToolName } from './schemas'
-export { LLM_STATIC_TOOLS } from './tools'
+export type { ClientToolName } from './tools'
+export { CLIENT_TOOL_NAMES, isClientToolName, LLM_STATIC_TOOLS } from './tools'
